Manual setting of redirect_uri in auth service
Currently you cannot change the redirect_uri when using OAuth2 or the old style of openidconnect configuration. As the file picker is to be used as parts of other projects and requires its own oidc-callback.html file to be redirected to I propose to make the redirect_uri configurable. The advantage is that the oidc-callback.html then can reside in any subfolder of the project.